Add Summon Immunity/Resistence, and greater improvements to cat summoning
Other Improvements for the summon system: Users are now able to select the level and form of the unit they want to summon. Currently most of this system works except for the filter and PCoins for summoned entities. There might be other changes not related to summon, but those are accidental (with the exception of the DPS column) so they can be discarded